Add websocket client to census. Create services for listening to census websocket. Created Voidwell.Cache.

+++ src/Voidwell.DaybreakGames/Models/MapScore.cs
@@ -5,8 +5,8 @@ namespace Voidwell.DaybreakGames.Models
     public class MapScore
     {
         public IEnumerable<int> Territories { get; set; }
-        public IEnumerable<double> Percent { get; set; }
+        public IEnumerable<float> Percent { get; set; }
         public IEnumerable<int> ConnectedTerritories { get; set; }
-        public IEnumerable<double> ConnectedPercent { get; set; }
+        public IEnumerable<float> ConnectedPercent { get; set; }
     }
 }
